requirements.txt
torch>=2.10.0
transformers>=5.2.2
sentence-transformers>=5.2.2
tokenizers>=0.22.2

langchain>=1.2.10
langchain-community>=0.4.1
langchain-core>=1.2.11
faiss-cpu>=1.13.2
langchain-huggingface>=1.2.0
langchain-groq>=1.1.2

numpy>=1.26.0
pandas>=2.2.0

pydantic>=2.7.0
typing-extensions>=4.12.0

PyPDF2>=3.0.1
python-docx>=1.2.0
pytesseract>=0.3.13